buttons merge
