+++ test/app_test.dart
@@ -2,7 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:memorizer/main.dart' as MAIN;
-import 'package:memorizer/modules/my_button.dart';
+import 'package:memorizer/modules/myButtons.dart';
+
+//import 'package:memorizer/modules/my_button.dart';
 import 'package:memorizer/modules/text_splitter_service.dart';
 
 void main() {
